requirements.txt
Django==1.9
djangorestframework==3.3.2
django-filter==0.11.0
psycopg2==2.6.1
